feat: complete challenge

+++ crud-restapi-challenge/src/Repositories/ItemRepository.cs
@@ -1,5 +1,6 @@
 using crud_restapi_challenge.Entities;
 using crud_restapi_challenge.Repositories.Interfaces;
+using crud_restapi_challenge.Exceptions;
 using Microsoft.EntityFrameworkCore;
 
 namespace crud_restapi_challenge.Repositories
@@ -31,15 +32,21 @@ namespace crud_restapi_challenge.Repositories
 
         public async Task UpdateItemAsync(Item item)
         {
-            _context.Entry(item).State = EntityState.Modified;
+            _context.Items.Update(item);
             await _context.SaveChangesAsync();
         }
 
         public async Task DeleteItemAsync(long id)
         {
             var item = await _context.Items.FindAsync(id);
+            if (item == null)
+            {
+                // Handle not found case, for example, throw an exception or return.
+                throw new NotFoundException($"Item with id {id} not found.");
+            }
             _context.Items.Remove(item);
             await _context.SaveChangesAsync();
         }
+
     }
 }
